ee-bench/packages/ee_bench_cli/pyproject.toml
[project]
name = "ee_bench_cli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ee_bench_generator",
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ee-dataset = "ee_bench_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ee_bench_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
